+++ bin/change_log.mli
@@ -4,55 +4,8 @@
    %%NAME%% %%VERSION%%
   ---------------------------------------------------------------------------*)
 
-open Cmdliner
-
-let cmds =
-  [
-    Tag.cmd;
-    Distrib.cmd;
-    Publish.cmd;
-    Opam.cmd;
-    Help.cmd;
-    Bistro.cmd;
-    Lint.cmd;
-    Check.cmd;
-    Delegate_info.cmd;
-    Config.cmd;
-    Undraft.cmd;
-    Change_log.cmd;
-  ]
-
-(* Command line interface *)
-
-let doc = "Release dune packages to opam"
-let sdocs = Manpage.s_common_options
-let exits = Cli.exits
-
-let man =
-  [
-    `S Manpage.s_description;
-    `P "$(mname) releases dune packages to opam.";
-    `P
-      "Without arguments, $(mname) acts like $(b,dune-release bistro): refer \
-       to $(b,dune-release help bistro) for help about the default behavior.";
-    `P "Use '$(mname) help release' for help to release a package.";
-    `Noblank;
-    `P "Use '$(mname) help troubleshoot' for a few troubleshooting tips.";
-    `Noblank;
-    `P "Use '$(mname) help $(i,COMMAND)' for help about $(i,COMMAND).";
-    `S Manpage.s_bugs;
-    `P "Report them, see $(i,%%PKG_HOMEPAGE%%) for contact information.";
-    `S Manpage.s_authors;
-    `P "Daniel C. Buenzli, $(i,http://erratique.ch)";
-  ]
-
-let main =
-  Cmd.group ~default:Bistro.term
-    (Cmd.info "dune-release" ~version:"%%VERSION%%" ~doc ~sdocs ~exits ~man)
-    cmds
-
-let main () = Stdlib.exit @@ Cmd.eval' main
-let () = main ()
+val cmd : int Cmdliner.Cmd.t
+(** The [change-log] command. *)
 
 (*---------------------------------------------------------------------------
    Copyright (c) 2016 Daniel C. Bünzli